Update config files: modularize BLK_DEV_MD, disable MD_AUTODETECT
Michal Marek says:
I think that the reason why BLK_DEV_MD is built in is

config MD_AUTODETECT
        bool "Autodetect RAID arrays during kernel boot"
        depends on BLK_DEV_MD=y

but since we compile the raid presonalities as modules, this cannot
work. Also, we use the initrd even in the trivial cases, so why bother
about root-on-md.

Neil Brown adds:
I certainly don't want customers using MD_AUTODETECT so I am quite
happy for md to become a module and MD_AUTODETECT to be disabled.

And Lars Marowsky-Bree adds:
It seems conceivable that the cluster RAID code might possibly be a
KMP which overrides basic md for a while. So making it a module is a
good idea.