- Update x86_64 config files: CONFIG_THERMAL=y
The generic thermal module is used by several ACPI components
(processor, fan, thermal and video) and a few other drivers. So
in practice it will end up being loaded on almost all systems
which support ACPI. It is small so build it into the kernel to
improve boot speed on x86_64.