supported.conf: Support the AD7879 touchscreen again
ad7879 was originally a standalone driver, that could be built to
support either I2C or SPI access. In kernel 2.6.36, it was split into a
helper module (ad7879) and two device drivers, one for I2C access
(ad7879-i2c) and one for SPI access (ad7879-spi, which we do not include
in SLE12.)

So in order to fully support the AD7879 touchscreen as we used to, we
must support both ad7879 and ad7879-i2c.