Update config files: re-enable twofish crypto support
Software twofish crypto support was disabled in several architectures
since SLES11 GA. For x86_64 it was on purpose, because
hardware-accelerated alternatives exist. However for some other
architectures it was by accident.

Additionally, in SLES11 SP2, the software twofish implementation was
renamed from twofish to twofish_generic, but supported.conf was not
updated accordingly, so S/390 was without a supported twofish crypto
implementation as well.

Re-enable software twofish crypto support in all configuration files,
to guarantee that at least one implementation is always available,
and adjust supported.conf to reflect the name change (bnc#871325).